Added Debug_Level parameter to key parameters for Zoltan. Still need to decide whether to standardize what specific Debug_Levels do or to let each algorithm decide.
Defined PARAM_VARS structures as static global variables in the files
that use them.  By making this change, only one definition
each set of parameters is needed.  The previous implementation
required two definitions of the parameters, making it more
likely that an error would be made when parameters were added
or deleted.