New_IDs_Branch merged with main trunk of Zoltan!
The data types for IDs in Zoltan are no longer compiled into the library.
Instead, the IDs are now arrays of unsigned integers.  The number of
integer entries in an ID can be set through LB_Set_Param.  Global and local
IDs may have different numbers of entries.  The LB_ID_TYPE is currently
defined to be unsigned integers; in the F90 interface, IDs are integers.

lb:
+ New ID formats defined in lbi_const.h.
+ Changes to support new IDs and argument lists for Zoltan functions and query
  functions.  Numbers of entries for global and local IDs are stored in the
  LB_Struct data structure.  All functions using global and/or local IDs now
  also include an argument for number of entries in the global and/or local IDs.
+ Removed LB_TAG data structure.
+ Removed use of LB_TAG in LB_Compute_Destinations; extra calls to LB_Comm_Do
  are used.
+ New files:     lb/lb_id.c, lb/lb_id_const.h
  Removed files: lb/lb_user_const.h

Still remaining to do:
+ Testing, testing, testing (any volunteers?).