New_IDs_Branch merged with main trunk of Zoltan!
The data types for IDs in Zoltan are no longer compiled into the library.
Instead, the IDs are now arrays of unsigned integers.  The number of
integer entries in an ID can be set through LB_Set_Param.  Global and local
IDs may have different numbers of entries.  The LB_ID_TYPE is currently
defined to be unsigned integers; in the F90 interface, IDs are integers.

rcb:
+ Changes to support new IDs and argument lists for Zoltan functions and query
  functions.
+ Combined some shared functionality of RCB and IRB into shared routines.
  In particular, many operations on Dots (building, communicating, etc.)
  are now shared.  (Less to maintain in the future.)   Other functionality
  could probably also be moved to these files and shared.
  New files:     rcb/shared.c, rcb/shared_const.h
+ In inertial2d.c, set cm[0] and cm[1] return arguments; I think they are used
  in the drops but they were never set.

Still remaining to do:
+ Testing, testing, testing (any volunteers?).