New_IDs_Branch merged with main trunk of Zoltan!
The data types for IDs in Zoltan are no longer compiled into the library.
Instead, the IDs are now arrays of unsigned integers.  The number of
integer entries in an ID can be set through LB_Set_Param.  Global and local
IDs may have different numbers of entries.  The LB_ID_TYPE is currently
defined to be unsigned integers; in the F90 interface, IDs are integers.

reftree:
+ Changes to support new IDs and argument lists for Zoltan functions and query
  functions.
+ Bill added quadrisection of quadrilaterals and fixed some bugs.
+ Changed communication using MPI_BYTE to use LB_ID_MPI_TYPE.
+ Replaced LB_GID data structure in LB_reftree_hash_node;
  IDs are now dynamically allocated for hash nodes.

Still remaining to do:
+ Testing, testing, testing (any volunteers?).