----------------------------------------------------------------------  Modified Files:  	build/makefile.LINUX.MPILAM  	packages/epetra/src/Epetra_BlockMap.cpp  	packages/epetra/src/Epetra_CrsGraph.cpp  	packages/epetra/src/Epetra_CrsGraph.h  	packages/epetra/src/Epetra_CrsMatrix.cpp  	packages/epetra/src/Epetra_MultiVector.cpp  	packages/epetra/src/Epetra_Object.cpp  	packages/epetra/src/Epetra_Object.h  	packages/epetra/src/Epetra_RowMatrix.h  	packages/epetra/src/Epetra_VbrMatrix.cpp  	packages/epetra/src/Epetra_Vector.cpp  	packages/epetra/src/Epetra_Vector.h  	packages/epetra/test/Makefile.template  	packages/epetra/test/BlockMap/cxx_main.cpp  	packages/epetra/test/CrsGraph/cxx_main.cpp  	packages/epetra/test/CrsMatrix/cxx_main.cpp  	packages/epetra/test/CrsRectMatrix/cxx_main.cpp  	packages/epetra/test/ImportExport/cxx_main.cpp  	packages/epetra/test/Map/cxx_main.cpp  	packages/epetra/test/MultiVector/BuildTestProblems.cpp  	packages/epetra/test/MultiVector/ExecuteTestProblems.cpp  	packages/epetra/test/MultiVector/ExecuteTestProblems.h  	packages/epetra/test/MultiVector/cxx_main.cpp  	packages/epetra/test/Object/cxx_main.cpp  	packages/epetra/test/SerialDense/cxx_main.cpp  	packages/epetra/test/SerialSpdDense/cxx_main.cpp  	packages/epetra/test/VbrMatrix/cxx_main.cpp  	packages/epetra/test/Vector/BuildTestProblems.cpp  	packages/epetra/test/Vector/BuildTestProblems.h  	packages/epetra/test/Vector/ExecuteTestProblems.cpp  	packages/epetra/test/Vector/ExecuteTestProblems.h  	packages/epetra/test/Vector/cxx_main.cpp  Added Files:  	packages/epetra/test/testAllMpi  ----------------------------------------------------------------------
The main purpose of this collection of mods is to make the automated testing results
for Epetra more meaningful and concise.  The primary feature is the ability to dynamically change
the traceback behavior.  In particular, when testing error code values (that is, when
putting in arguments that should cause an error flag to be generated), it is now
possible to shut of the error traceback output completely.  The traceback information was
very annoying and confusing.

To accomplish this we now have a static int and several static methods
in the Epetra_Object class to control the traceback macro behavior.  Since every
other Epetra class derives from Epetra_Object, the traceback control can be
executed via SetTracebackMode() using any Epetra object instance and it will
affect all objects.

In addition to this feature, there are numerous other small bugfixes.

There is an automated script (testAllMpi) for running MPI versions of all
Epetra unit tests.  It is specific to my Linux environment, but it is trivial to
modify it.

I ran all regression tests in both serial and MPI mode.  All tests work, and the
results are far easier to understand (I think).