Added "reset" capability to Solvers and Linesearches. This allows us to reuse the same solver over and over again and is useful, e.g., when a nonlinear solver is called every time step of an ODE/PDE solver.
Also added "Max Iters" parameter to all the line searches (except
FullStep).

Also changed the output of each line search to include the iteration
number.

Also changed the output of the Newton iteration to be shorter.

Also added the ability for the Solver::Manager to change solvers when
it is reset.