----------------------------------------------------------------------  Committing in packages
 Modified Files:
 	aztecoo/example/AztecOO/cxx_main.cpp aztecoo/src/AztecOO.cpp
 	epetra/src/Epetra_CrsMatrix.cpp epetra/src/Epetra_CrsMatrix.h
 	epetra/src/Epetra_MultiVector.cpp
 	epetra/src/Epetra_MultiVector.h epetra/src/Epetra_Object.h
 	epetra/src/Epetra_Operator.h epetra/src/Epetra_VbrMatrix.h
 	ifpack/src/Ifpack_CrsRiluk.cpp ifpack/src/Ifpack_CrsRiluk.h
 	komplex/example/komplex_hb/main.c
 	triutils/src/Trilinos_Util_ReadHb2Epetra.cpp
 Added Files:
 	aztecoo/example/IfpackAztecOO/Makefile
 	aztecoo/example/IfpackAztecOO/cxx_main.cpp
 ----------------------------------------------------------------------

Made a number of small changes due to bugs and desired features for solving
transpose systems.  AztecOO now works (has been tested) solving the transpose problem
of the matrix that is passed in.  Also, Ifpack has been tested as a preconditioner for AztecOO
and can be used as the preconditioner for the transpose problem.

I made one addition to the Epetra_Operator header file by adding a Label() method.  This allows
a descriptive name to be associated with operators.

I have also added a new example driver to illustrate the use of Ifpack with AztecOO.

I ran all test and sample codes for epetra, ifpack and aztecoo.  Everything work in serial and parallel.