----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	packages/epetra/src/Epetra_BlockMap.cpp
 	packages/epetra/test/BlockMap/checkmap.cpp
 	packages/epetra/test/BlockMap/cxx_main.cpp
 Added Files:
 	packages/epetra/test/ReferenceLogs/log.12Mar2002:15:37:36
 	packages/epetra/test/ReferenceLogs/log.mpi.12Mar2002:15:46:02
 ----------------------------------------------------------------------
Added a test for the ElementSize(int LID) method.

In the process found and fixed a bug that prevented Epetra from working properly in a serial environment when the GID's were not assigned values sequentially from IndexBase to NumMyElements-1+IndexBase.  This fix involved changing LID(int GID), GID(int LID), and GlobalToLocalSetup.  Thanks to Mike H for pointing out that the problem extended to the GlobalToLocalSetup function.