----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	test/testAll test/testAllMpi test/BlockMap/checkmap.cpp
 	test/BlockMap/checkmap.h test/BlockMap/cxx_main.cpp
 	test/Comm/cxx_main.cpp test/CrsGraph/cxx_main.cpp
 	test/CrsMatrix/cxx_main.cpp test/CrsRectMatrix/cxx_main.cpp
 	test/ImportExport/cxx_main.cpp test/Map/checkmap.cpp
 	test/Map/checkmap.h test/Map/cxx_main.cpp
 	test/MultiVector/ExecuteTestProblems.cpp
 	test/MultiVector/ExecuteTestProblems.h
 	test/MultiVector/cxx_main.cpp test/Object/cxx_main.cpp
 	test/VbrMatrix/cxx_main.cpp
 	test/Vector/ExecuteTestProblems.cpp
 	test/Vector/ExecuteTestProblems.h test/Vector/cxx_main.cpp
 Added Files:
 	test/epetra_test_err.h
 ----------------------------------------------------------------------

This update incorporates a couple of changes into the test suite.

First, the tests for each class were upgraded so that after an error occurs, the rest of the tests are still completed, and any more errors that are found are also reported.  The old standard used assert commands, which omitted the remaining tests.  The new file epetra_test_err.h includes a macro that is used in place of assert.

Second, the testAll scripts were improved again to allow for better automated testing and to restore the old default functionality - users once again do not have to pass any parameters, default parameters are supplied (TRILINOS_ID=_DEBUG is used for both serial and mpi).