Made a trivial change to Epetra_Object.h:   Changed "std::sprintf" to "sprintf". janus doesn't have <cstdio>, and   sprintf isn't in the std:: namespace on janus. For other platforms that   do put sprintf in std::, Epetra_Object.h already has 'using namespace std;'   statements.
Modified several of the esi adaptors so that they no longer inherit from
the Epetra classes that they use to implement the ESI interfaces. Instead,
they now simply hold an instance of the Epetra class as member data, and
make calls to that instance. Breaking the inheritance chain solved some
run-time problems that were occuring in code that uses the esi stuff.