Changed PrecMatrix to Preconditioner throughout.
Added new function applyJacobianInverse.

Changed some of the functions from pure virtual to virtual - this is
to allow simple defaults for functions that might not be supported in
derived objects. This will also be useful as we add additional
functionality to the Group.

Added new file NOX_Abstract_Group.C to define defaults for the virtual
functions.