Added an enumerated type to the computeF() function in NOX::Epetra::Interface.
The function previously was:

virtual bool computeF(const Epetra_Vector& x, Epetra_Vector& RHS) = 0;

and now looks like:

virtual bool computeF(const Epetra_Vector& x, Epetra_Vector& RHS,
                      FillType flag = F) = 0;

Where the enumerated type can be "F" or "Jacobian".  This additional term
is used to tell the user's code why computeF() is being called.  In some
cases, if we are calling computeF() to be used to compute an estimate to the
Jacobian (i.e. generating the Jacobian via finite differencing) we may not
want to include all of the terms in the Jacobian.  This allows the user to
"lag" certain derivatives in a finite difference or matrix-free Jacobian.
This is critical in Radiation transport in ALEGRA and has been used in
the past in MPSalsa.  It has been observed that a more exact Jacobian can
in some instances lead to poorer performance of the linear solver.