----------------------------------------------------------------------
 Modified Files:
 	az_tools.c
 ----------------------------------------------------------------------

Small change to AZ_find_procs_for_externs in az_tools.c.  This bug was
found by Jonathan Hu and Allen Robinson.  It appears that the bug has been
around for a long time and is the source for potentially many memory growth
problems in Aztec.  In particular, this could fix a problem that Laurie Frink
was seeing with memory use.

Mike