Re-fixed the MergeRedundantEntries bug in Epetra_CrsMatrix after finding another case that it wasn't handling correctly.
Added initial implementation of Epetra_FECrsMatrix. This is a class that
inherits Epetra_CrsMatrix and adds the ability to accept finite-element
style sub-matrices as input. These sub-matrices need not be wholly locally-
owned. For example, they may contain data corresponding to shared finite-
element nodes in a parallel context. There is a GlobalAssemble() method that
should be called after data input is complete, to perform the communication
needed to gather all non-local (shared) data onto the owning processors as
determined by the map that the matrix was constructed with.