Additions to the NOX::Epetra::FiniteDifference object to better handle poorly scaled problems.  Such problems were encountered in the Sierra/Premo code last week.  Uknowns had a difference in magnitude on the order of 1.0e5 and under certain initial guesses the estimated perturbation parameter was negligible. Two options have been added:
1. The application can supply a vector of the size of the number of unknowns
that conatains the average values of the unknowns so that the perturbation
should never be negligible (see the doxygen documentation for more).

2. The application can now supply an Epetra_CrsGraph so the FiniteDifference
object does not have to attempt to determine the graph.

These changes should not affect any code already using the FiniteDifference
objects since extra constructors were added for these options.