Sample problem to test finite diferencing using coloring.
The coloring algorithm is the greedy algorithm implemented by Bill Spotz and
adapted to Epetra by Rob Hoekstra.  The algorithm resides in the
contrib/epetra_transform directory in Trilinos, and the required files have
been copied here to make a stand-alone test problem.  Extensions of the
copied files have been changed from .cpp to .C and .h to .H in keeping with
the NOX standard.

The example is based on epetra-examples/1Dfem_nonlinear but allows
creation of a NOX::Epetra::FiniteDifferenceColoring object.  The user is
required to provide an Epetra_MapColoring object and a vector<Epetra_IntVector>
object in the constructor.  The latter provides column indices where computed
values are to placed in the Jacobian for a given row and color.
For now, creation of the color map and the column index map is limited to
serial execution.