----------------------------------------------------------------------
 Added Files:
 	Epetra_LinearProblemRedistor.cpp
 	Epetra_LinearProblemRedistor.h
 ----------------------------------------------------------------------

A new class to support distributed direct solvers.  This class takes an
existing distributed Epetra_LinearProblem and redistributes it, either using
a user-defined target map or by constructing a map such that PE 0 owns an
entire copy of the matrix (and optionally all PEs can own a copy).

This class will also transpose the matrix if requested, and will make the
matrix and multivector data contiguous and compatible with Fortran array storage.

These files are works in progress, but wanted to check them in for safety.